format_issues/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
deprecated==1.2.18
google-auth==2.40.2
google-genai==1.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pygithub==2.6.1
pyjwt==2.10.1
pynacl==1.5.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
websockets==15.0.1
wrapt==1.17.2
